Organize repo structure and add comprehensive README
