+++ .gitignore
@@ -1,0 +1,7 @@
+# Byte-compiled files
+__pycache__/
+*.pyc
+
+# MacOS metadata
+.DS_Store
+
