+++ examples/sample_ndvi2.png
