+++ examples/sample_ndvi3.png
