+++ examples/sample_ndvi4.png
